requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
firebase-admin==6.4.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
gunicorn==21.2.0
eventlet==0.33.3
